requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
dotenv==0.9.9
idna==3.10
python-dotenv==1.1.0
requests==2.32.3
urllib3==2.3.0
gspread
google-auth
